Performed first test

+++ src/videocomposing.py
@@ -4,6 +4,7 @@ import pdb
 import midiparse
 import random
 import os
+import audioanalysis
 
 
 def compose(instruments, midipattern, width, height, source_dir):
@@ -15,25 +16,36 @@ def compose(instruments, midipattern, width, height, source_dir):
     for i, track in enumerate(midipattern[1:]):
         print "Composing track " + str(i) + "..."
         name = midiparse.get_instrument_name(track)
+        file_name = name + '.mp4'
+        if os.path.isfile(file_name):
+            written_clips.append((len(track), file_name))
+            continue
         if name is None:
             # FIXME this is ugly
             name = "Untitled Instrument 1"
         instrument_clips = _load_instrument_clips(name, 
                                                   instruments[name],
                                                   source_dir)
-        track_clip = _process_track(instrument_clips,
+        try:
+            track_clip = _process_track(instrument_clips,
                                           track, pulse_length, width, height)
-        file_name = name + '.mp4'
-        track_clip.write_videofile(file_name)
-        written_clips.append(file_name)
-        _delete_clips(instrument_clips)
+            track_clip.write_videofile(file_name)
+            written_clips.append((len(track), file_name))
+            _delete_clips(instrument_clips)
+            del instrument_clips
+        except Exception as e:
+            print "Couldn't process instrument {}: {}, continuing...".format(
+                name, e.message)
+            continue
+
+    written_clips.sort(key=lambda s: s[0], reverse=True)
 
     final_clips = []
-    for i, file_name in enumerate(written_clips):
+    for i, (_, file_name) in enumerate(written_clips):
         clip = edit.VideoFileClip(file_name)
-        x, y, w, h = _partition(width, height, len(track_clips), i)
+        x, y, w, h = _partition(width, height, len(written_clips), i)
         final_clips.append(
-            fx.resize(clip, newsize=(w, h)).set_position(x, y))
+            fx.resize(clip, newsize=(w, h)).set_position((x, y)))
     return edit.CompositeVideoClip(size=(width, height), clips=final_clips)
 
 
@@ -70,25 +82,35 @@ def _partition(width, height, num_sim_notes, pos):
         return 0, 0, width, height
     w = width//2
     h = height//2
-    if pos < 4:
-        # put it in one of 4 spots
-        x = (pos // 2)*w
-        y = (pos % 2)*h
+    if num_sim_notes <= 4:
+        if pos < 4:
+            # put it in one of 4 spots
+            x = (pos // 2)*w
+            y = (pos % 2)*h
 
-        if num_sim_notes < 2:
-            # if there are only two notes, vertically center
-            # the clips 
-            y += h
-        elif num_sim_notes == 3 and pos == 2:
-            # if there are a total of 3 clips, horizontally
-            # center the last clip
-            x += w
-        return x, y, w, h
+            if num_sim_notes < 2:
+                # if there are only two notes, vertically center
+                # the clips 
+                y += h
+            elif num_sim_notes == 3 and pos == 2:
+                # if there are a total of 3 clips, horizontally
+                # center the last clip
+                x += w
+            return x, y, w, h
+        else:
+            # if there are more than 4, place it randomly on top of the others
+            rx = random.randint(-w//2, w//2) + w
+            ry = random.randint(-h//2, h//2) + h
+            return rx, ry, w, h
     else:
-        # if there are more than 4, place it randomly on top of the others
-        rx = random.randint(-w//2, w//2) + w
-        ry = random.randint(-h//2, h//2) + h
-        return rx, ry, w, h
+        w = width//3
+        h = height//3
+        if pos > 10:
+            # place it off screen
+            return width, height, w, h
+        return (pos // 3)*w, (pos % 3)*h, w, h
+
+
 
 
 def _process_track(clips, midi_track, pulse_length, width, height):
@@ -106,9 +128,11 @@ def _process_track(clips, midi_track, pulse_length, width, height):
         num_sim_notes = curr_event.num_simultaneous_notes
         x, y, w, h = _partition(width, height, 
                                 num_sim_notes, note.video_position)
+        #if note.duration*pulse_length*2 < audioanalysis.STANDARD_OFFSET:
+        clip = clip.subclip(audioanalysis.STANDARD_OFFSET)
         clip = clip.set_start(note.start*pulse_length)
         d = clip.duration
-        clip = clip.set_duration(min(note.duration*pulse_length, d))
+        clip = clip.set_duration(min(note.duration*pulse_length*2, d))
         clip = clip.set_position((x, y))
         clip = fx.resize(clip, newsize=(w, h))
         parsed_clips.append(clip)
